unified tests files
Now both rococo and westend call the same test files
